requirements.txt
chromadb==0.4.15
gradio==4.19.1
pyautogen==0.2.0
openai==1.3.6
pypdf==3.17.0
IPython==8.17.2
sentence_transformers==2.2.2

